Modularize pfcat/pfgrep
A single file now handles most of the file iteration and reading,
and pfgrep/pfcat.c just contain the command line parsing, prep, and
actual work on the read-in files. This is made possible by the fact
we basically convert record files into "normal" streamfiles.
