+++ .gitignore
@@ -1,3 +1,4 @@
 *.o
 pfgrep
+pfcat
 core
